move make_msg to helper file
